Adds per-game config when loading a game from args
Per-game config can only be loaded after the CD image is loaded, which
has to happen after plugins are initialized, which has to happen after
the config is loaded. This change loads the global config first, and
tries to load per-game config after the CD is loaded. This may not
work 100% of the time, but works better than it does right now.