Adds an option for scaling 256px-wide screens
This option takes over the Software Scaling option in the display
menu. There is a performance cost (about 10%). It does subpixel scaling
to keep the image as sharp as it can be.

This also moves scaling into its own loop -- I saw a 4-6% increase in
performance when upscaling and downscaling after making that change.